larger refactor to allow for other menus, unrelated to mod options, to be registered
