+++ UnboundLib/Unbound.cs
@@ -107,7 +107,7 @@ namespace UnboundLib
                     text.rectTransform.localPosition = new Vector3(0, 325, text.rectTransform.localPosition.z);
                 });
 
-                UI.MenuHandler.Instance.CreateModOptions(firstTime);
+                UI.ModOptions.Instance.CreateModOptions(firstTime);
                 
                 firstTime = false;
 
@@ -225,24 +225,24 @@ namespace UnboundLib
 
         private void Update()
         {
-            if (Input.GetKeyDown(KeyCode.F1) && !UI.MenuHandler.noDeprecatedMods)
+            if (Input.GetKeyDown(KeyCode.F1) && !UI.ModOptions.noDeprecatedMods)
             {
-                UI.MenuHandler.showModUi = !UI.MenuHandler.showModUi;
+                UI.ModOptions.showModUi = !UI.ModOptions.showModUi;
             }
 
-            GameManager.lockInput = UI.MenuHandler.showModUi || DevConsole.isTyping;
+            GameManager.lockInput = UI.ModOptions.showModUi || DevConsole.isTyping;
         }
 
         private void OnGUI()
         {
-            if (!UI.MenuHandler.showModUi) return;
+            if (!UI.ModOptions.showModUi) return;
 
             GUILayout.BeginVertical();
 
             bool showingSpecificMod = false;
-            foreach (var md in UI.MenuHandler.GUIListeners.Keys)
+            foreach (var md in UI.ModOptions.GUIListeners.Keys)
             {
-                var data = UI.MenuHandler.GUIListeners[md];
+                var data = UI.ModOptions.GUIListeners[md];
                 if (data.guiEnabled)
                 {
                     if (GUILayout.Button("<- Back"))
@@ -265,9 +265,9 @@ namespace UnboundLib
             }
 
             GUILayout.Label("Mod Options:");
-            foreach (var md in UI.MenuHandler.GUIListeners.Keys)
+            foreach (var md in UI.ModOptions.GUIListeners.Keys)
             {
-                var data = UI.MenuHandler.GUIListeners[md];
+                var data = UI.ModOptions.GUIListeners[md];
                 if (GUILayout.Button(data.modName))
                 {
                     data.guiEnabled = true;
@@ -376,12 +376,12 @@ namespace UnboundLib
 
         public static void RegisterMenu(string name, UnityAction buttonAction, Action<GameObject> guiAction, GameObject parent = null)
         {
-            UI.MenuHandler.Instance.RegisterMenu(name, buttonAction, guiAction, parent);
+            UI.ModOptions.Instance.RegisterMenu(name, buttonAction, guiAction, parent);
         }
 
         public static void RegisterGUI(string modName, Action guiAction)
         {
-            UI.MenuHandler.RegisterGUI(modName, guiAction);
+            UI.ModOptions.RegisterGUI(modName, guiAction);
         }
         public static void RegisterHandshake(string modId, Action callback)
         {
